Crate hyperon_space

Space is a storage for atoms with search queries execution capability. This module is intended to keep different space implementations.

Modules

index

Structs

DynSpace

SpaceCommon

A common object that needs to be maintained by all objects implementing the Space trait

SpaceObserverRef

A reference to a SpaceObserver that has been registered with a Space

Enums

SpaceEvent

Contains information about space modification event.

Constants

ATOM_TYPE_SPACE

COMMA_SYMBOL

Symbol to concatenate queries to space.

Traits

Space

Read-only space trait.

SpaceMut

Mutable space trait.

SpaceObserver

Space modification event observer trait.

SpaceVisitor

An interface for visiting space atoms.

Functions

complex_query