hyperon_space

Trait Space

pub trait Space: Debug + Display {
    // Required methods
    fn common(&self) -> FlexRef<'_, SpaceCommon>;
    fn query(&self, query: &Atom) -> BindingsSet;
    fn visit(&self, v: &mut dyn SpaceVisitor) -> Result<(), ()>;
    fn as_any(&self) -> &dyn Any;

    // Provided methods
    fn subst(&self, pattern: &Atom, template: &Atom) -> Vec<Atom> { ... }
    fn atom_count(&self) -> Option<usize> { ... }
}

Read-only space trait.

Required Methods

fn common(&self) -> FlexRef<'_, SpaceCommon>

Access the SpaceCommon object owned by the Space

fn query(&self, query: &Atom) -> BindingsSet

Executes query on the space and returns variable bindings found. Query may include sub-queries glued by [grounding::COMMA_SYMBOL] symbol. Each Bindings instance in the returned [BindingsSet] represents single result.

Examples

use hyperon_atom::{expr, bind_set, sym};
use hyperon_atom::matcher::BindingsSet;
use hyperon::space::grounding::GroundingSpace;

let space = GroundingSpace::from_vec(vec![expr!("A" "B"), expr!("B" "C")]);
let query = expr!("," ("A" x) (x "C"));

let result = space.query(&query);

assert_eq!(result, bind_set![{x: sym!("B")}]);

fn visit(&self, v: &mut dyn SpaceVisitor) -> Result<(), ()>

Visit each atom of the space and call SpaceVisitor::accept method. This method is optional. Return Err(()) if method is not implemented. Cow<Atom> is used to allow passing both references and values. First is appropriate for collection based atomspace. Second one is more usable if atomspace is a generator or when values cannot be extracted easily and should be reconstructed instead.

fn as_any(&self) -> &dyn Any

Returns an &dyn Any for spaces where this is possible

Provided Methods

fn subst(&self, pattern: &Atom, template: &Atom) -> Vec<Atom>

Executes pattern query on the space and for each result substitutes variables in template by the values from pattern. Returns results of the substitution.

Examples

use hyperon_common::assert_eq_no_order;
use hyperon_atom::expr;
use hyperon_space::Space;
use hyperon::space::grounding::GroundingSpace;

let space = GroundingSpace::from_vec(vec![expr!("A" "B"), expr!("A" "C")]);

let result = space.subst(&expr!("A" x), &expr!("D" x));

assert_eq_no_order!(result, vec![expr!("D" "B"), expr!("D" "C")]);

fn atom_count(&self) -> Option<usize>

Returns the number of Atoms in the space, or None if this can’t be determined