hyperon_space

Trait SpaceMut

pub trait SpaceMut: Space {
    // Required methods
    fn add(&mut self, atom: Atom);
    fn remove(&mut self, atom: &Atom) -> bool;
    fn replace(&mut self, from: &Atom, to: Atom) -> bool;
    fn as_any_mut(&mut self) -> &mut dyn Any;
}

Mutable space trait.

Required Methods

fn add(&mut self, atom: Atom)

Adds atom into space.

Examples

use hyperon_atom::sym;
use hyperon_atom::matcher::BindingsSet;
use hyperon::space::grounding::GroundingSpace;

let mut space = GroundingSpace::from_vec(vec![sym!("A")]);
 
space.add(sym!("B"));

assert_eq!(space.query(&sym!("A")), BindingsSet::single());
assert_eq!(space.query(&sym!("B")), BindingsSet::single());
assert_eq!(space.query(&sym!("C")), BindingsSet::empty());

fn remove(&mut self, atom: &Atom) -> bool

Removes atom from space. Returns true if atom was found and removed, and false otherwise.

Examples

use hyperon_atom::sym;
use hyperon_atom::matcher::BindingsSet;
use hyperon::space::grounding::GroundingSpace;

let mut space = GroundingSpace::from_vec(vec![sym!("A")]);
 
space.remove(&sym!("A"));

assert_eq!(space.query(&sym!("A")), BindingsSet::empty());

fn replace(&mut self, from: &Atom, to: Atom) -> bool

Replaces from atom to to atom inside space. Doesn’t add to when from is not found. Returns true if atom was found and replaced, and false otherwise.

Examples

use hyperon_atom::sym;
use hyperon_atom::matcher::BindingsSet;
use hyperon::space::grounding::GroundingSpace;

let mut space = GroundingSpace::from_vec(vec![sym!("A")]);
 
space.replace(&sym!("A"), sym!("B"));

assert_eq!(space.query(&sym!("A")), BindingsSet::empty());
assert_eq!(space.query(&sym!("B")), BindingsSet::single());

fn as_any_mut(&mut self) -> &mut dyn Any

Returns an &mut dyn Any for spaces where this is possible